refactor: move contextprovider
